updated capital country.c
Aligned all the if-else statements in proper syntax. Fixed an error that used to occur earlier if a country name starting with "E" was entered, the loop used to break. Fixed it by removing that break statement. Fixed the spelling of SRI LANKA. Added proper line break after the "Don't Know" statement. Did the first letter of "Capital" as uppercase to make the output look better.